pyproject.toml
[project]
name = "ComfyUI-SAM3DBody"
description = "SAM 3D Body integration for ComfyUI - Robust full-body human mesh recovery from single images. Reconstruct 3D human meshes with pose, shape, and hand details. First draft, please open a github issue if you have any problems or feature requests :)"
version = "0.0.3"
license = { text = "MIT License (wrapper) + SAM License (vendored library)" }
readme = "README.md"
authors = [
    { name = "Andrea Pozzetti", email = "pznodes@gmail.com" }
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "opencv-python>=4.8.0",
    "pytorch-lightning>=2.0.0",
    "pyrender>=0.1.45",
    "yacs>=0.1.8",
    "scikit-image>=0.19.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "dill>=0.3.6",
    "pandas>=1.5.0",
    "rich>=13.0.0",
    "hydra-core>=1.3.0",
    "hydra-submitit-launcher>=1.2.0",
    "hydra-colorlog>=1.2.0",
    "pyrootutils>=1.0.4",
    "webdataset>=0.2.0",
    "chump>=1.6.0",
    "networkx==3.2.1",
    "roma>=1.3.0",
    "joblib>=1.2.0",
    "seaborn>=0.12.0",
    "wandb>=0.15.0",
    "appdirs>=1.4.4",
    "appnope>=0.1.3",
    "ffmpeg-python>=0.2.0",
    "cython>=0.29.0",
    "jsonlines>=3.1.0",
    "pytest>=7.4.0",
    "xtcocotools>=1.14",
    "loguru>=0.7.0",
    "optree>=0.9.0",
    "fvcore>=0.1.5",
    "black>=23.0.0",
    "pycocotools>=2.0.6",
    "tensorboard>=2.13.0",
    "huggingface-hub>=0.16.0",
    "transformers>=4.30.0"
]

[project.urls]
Repository = "https://github.com/PozzettiAndrea/ComfyUI-SAM3DBody"
Documentation = "https://github.com/PozzettiAndrea/ComfyUI-SAM3DBody/blob/main/README.md"

[tool.comfy]
PublisherId = "pznodes"
DisplayName = "ComfyUI-SAM3DBody"
Icon = ""

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests (fast, no model loading)",
    "integration: Integration tests with real models (slow, requires HF_TOKEN)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["nodes", "sam_3d_body"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
